requirements.txt
streamlit
fastapi
uvicorn[standard]
scikit-learn
umap-learn
hdbscan
sentence-transformers
openai
python-dotenv
supabase
reportlab
requests
pandas
numpy
plotly
streamlit-authenticator
streamlit-option-menu
openpyxl
xlrd
PyPDF2
pdfplumber
python-docx
